feat(frontend): integrate BlogApi
